Add support for derived_events __EMIT__
Previously derived events controlled 'when to write out a new
event' by the 'trigger_event'. When the derived event code saw
the trigger event then it was time to push the new event.
This worked okay for simple ftrace events which had a first and
last event (like a write start and a write end).
For a lot of other events, you want to write the new event when
you have a set complete. In particular, if you've collected a
stream of events (like perf -e "{cpu-clock,cycles,instructions}:S"
... a group of events with a group leader) then, when the leader
(cpu-clock) is triggered then all the events in the group are
collected. All the events will have the same time stamp and
duration. This lets OPPAT compute ratio (like CPI cycles/instr)
more reliably.
Adding the __EMIT__ 'system field name' allows the lua derived
event routine to control whether to write out an event. If the
__EMIT__ variable is set to 1 then the new event is assumed
complete and pushed onto the event vector.
See the CPI event in charts.json and src_lua/prf_CPI.lua file for
examples.