Add updated sample html files
The sample html files are all replaced with versions that use
zlib to compress the data. This will slow loading but greatly
reduces the size of the file.
Added msgs which display during the loading so the user isn't
completely befuddled. I don't compress the data for 'live'
displays, just for the 'standalone' html files.