Reduce setTimeout delays and server sleep delay
The setTimeout delays for parsing/charting data from the webserver
were too big and slowed down loading the web page.
The webserver sleep delay (in sending data to the browser) was
also too long and slowing down page loading.
Reduce the setTimeout to the minimum (1 ms) still allows the
page loading status messages to get updated. The webserver delay
was reduced to 200 ms.