Add ETW Marker chart
The chart just shows a marker indicating when the spin.cpp program
emitted a Marker (via EtwSetMark()).
This chart's flds use the new TYP_ADD_2_EXTRA qualifier.
TYP_ADD_2_EXTRA indicates that the text of the field should be
passed to the browser for display. Also the duration is set to
1.0e-6 seconds (because it doesn't really have a duration...
The event is just a point in time.
Also new is the chart marker field:
"marker":{"type":"square", "size":"10", "connect":"n"}
This indicated that squares should be draw of size 10 pixels and
lines should not be drawn connecting the points.