use fte_enum, add marker support
Most of the changes are due to changing the FLD_TYP_* enum to
a class with a namespace. the enum is a flag structure and it
reached 32 bits so it had to be moved to a 64bit field.
Also added code to send the marker fields (from charts.json) to
the browser.
Lastly, the ultimate goal of all this was to be able to display
user defined events and text such as a marker at T=x with the
text that the user specified, such as 'Begin mem_bw on thread 0'.
For ETW data, I needed a new field to to indicate that 'the text
in this field needs to be passed to the browser for display when
you hover over the marker'. Whereas perf data already passed text
to the browser, ETW didn't have this implemented.