Add support for markers and fix lkup
Add a draw_rect() routine to actually draw the markers.
Currently only 'square' markers are supported but the markers can
be different colors.
Also add support for highlighting in stacked charts.
The 'marker' chart is implemented as a line chart with markers
where the lines between points are not drawn... since there
really isn't a 'duration' for the events. The marker is just
telling us that an event X occurred at time Y with text 'Z'.
Add support for line charts without connecting lines (see the
charts.json chart.marker.connect=n fields.
Also fixed a problem where the search for a hit on a chart line
might fail if the events on the line overlapped on the X access.
If the chart line data overlaps then a linear search is used to
find which event the user is hovering over.