Add support for ftrace:print events
There are default events that don't get listed by 'perf list' nor
'trace-cmd list'.
These are in /sys/kernel/debug/tracing/events/ftrace/*
The trace_markers show up in the ftrace:print event.
trace-cmd automagically displays the events. I don't know how to
get perf to record/display the events.
But the event shows up without an area (ftrace:) in the file
so I have to add code to support checking to see if the 'print'
event is the ftrace:print event.
If it is then return the perf_event_dump_list ID number for the
event.