change 'zoom all' to just zoom all one time
Previously the 'zoom all' option in the left hand side menu put
the charts in a 'whenever 1 chart is zoomed/panned then zoom/pan
all the rest of the charts' mode. This led to weird lags where
the other graphs would be shifting while I was still shifting the
main chart.
Now when you click the 'zoom/pan all' option, the last 'zoom/pan'
is applied to all the other charts. So it is a one-time zoom/pan
of the other charts. If you then make another zoom/pan to a chart
and you want it applied to all the other charts, you have to hit
the 'zoom all' left-hand-side menu option again.