Add support for event multiplexing
Multiplex events present a challenge for plotting.
There may be gaps in the data (as groups are multiplexed in/out),
or the elapsed time (as measured by the time between the previous
sample and the current sample) may not agree with the
'time running' field (if present).
With this patch a TYP_TM_RUN charts.json event field flag is used
to indicate that the 'time running' field should be used as the
elapsed time for the sample (as opposed to the time between
samples).
Also add a charts.json charts options flag to drop the first point
on line charts if we aren't using the TYP_TM_RUN. In particular
this option is useful for when the elapsed time is the difference
in time (per cpu) for samples. On the 1st event, the elapsed time
is set to the current sample time - start time. This can result
in wildly short elapsed times for the 1st sample.
The DROP_1ST option just drops the 1st point on line charts.
Also generalized the reading/processing of options.