Handle multiplexing hardware event groups
To more accurately handle perf multiplexing of hardware events I
need the 'time running' and 'time enabled' field in the file.
This lets me know the exact interval covered by the event.
For example, say we are looking at the instructions event and it
is used in every multiplex group. Previously I would just say that
the interval for the instructions event was the difference between
the current instruction event timestamp and the previous instr
timestamp (for a given cpu). With multiplexing this gets more
complicated... the previous hw event might be 2 or groups previous
so the timestamp of the current group - prev group isn't the right
interval.
Also, on ARM raspberry pi b+ cortex A53 at least, the timestamps
seem to be not right. In particular when perf switches between 2
groups the timestamp can be too small... much shorter (10x?) than
the delta in 'time running'.
So we know record the delta in time running (if present) and use
it for the interval if the charts.json event fld TYP_TM_RUN is
used to request using 'running time' for the duration.
Also added code to differentiate which multiplex group is being
processed.