Add better clipping and clip to markers
The clipping (restricting the time to x-y) support has been pretty
crude. Having added support for markers, it is useful to be able
to clip relative to a marker.
Add --marker_beg_num and --marker_end_num options for clipping
relative a marker number. Eventually I should probably allow
selecting a marker based on the text in the marker rather than
a marker number. The list of markers (and number) is printed
in the stdout.
If you enter the --marker_beg_num/end_num then the -b/-e times
are relative to the marker's absolute time. Without the
--marker_beg/end_num options then you have to enter the abs time
for the -b/-e time.
The clipping is much improved now and (as far as I can tell) if
the chart has data extending out of the clip region then that data
will not be drawn.
The clip to a marker allows me to avoid knowing the absolute
timestamp (so when I update the data files I can regenerate the
sample files much more easily).