Add raw events to event groupings
I don't know what is going on but event grouping (with leaders)
didn't work properly on Intel. I want cpu-clock with a freq=997
to control when events are read. This gives me a read of the
events in the group about every 1 ms. Events like uops_issued.any
did not work in the group... they forced all the events to be
sampled using separate 'sampleafter' values.
I can workaround this by specifying the raw event encoding.
But this has the unwanted effect of requiring the charts.json
event/chart to use the raw encoding... reducing readability and
causing conflicts between architectures (since different arch's
can have the same raw encodings).