Add cpu_diagram charts and features
For the cpu_diagram, I need to be able to map a chart's data to
some resource on a cpu block diagram. The tot_line feature is
expanded to handle mapping cpus to cores (for HT enabled) and
a 'per chart' number has to be produced (like average memory bw)
for the current interval.
Also lots of charts are 'X event per cycle' or 'cycles per X
event' type of ratios. These ratios don't add well... the avg of
a bunch of ratios does not equal (sum of numerator)/(sum of
denominator). So I had to pass the numerator and denom of each
value to the web so the tot_line could be accurately calculated.
Over 50 charts are added for the cpu-diagram charts. Not all the
charts are used currently by the run_perf_x86_haswell.sh script.