Add support for creating numerator demonitor fields
Lots of the cpu_diagram charts need tot_lines and the tot_lines
values are ratios of 2 events (like cycles/uop).
The tot_line (in main.js) divides the selected chart interval into
100 subintervals and computes an average for the subinterval.
If the values in the subinterval are ratios then the average of
the values won't (usually) be equal to 'sum of the numerator'/'sum
of the denominator'.
Added the logic to emit the numerator and denominator fields for
the computed ratio values.