Generate the cpu_diagram for standalone html
The standalone version of the html files weren't showing the
cpu_diagram metrics on the canvas cpu_diagram. Also the metric
table wasn't generating.
The issue is that the metrics require the underlying charts to be
generated (which generates the metrics). So parse_svg() needs to
be called after all the charts are generated.
This change adds a step to standaloneJob() to wait until all the
charts are generated before calling parse_svg().
And I comment out a few console.log statements.