Allow multiple tests to be run using input file
Currently only 1 test can be run per invocation.
I wanted to be able to run a memory bw test, then an L3 bw test,
then an L2 test, then a spin test to show how the cpu_diagram
data changes.
So I added an input file where each line is a separate cmdline
options list.
If the 1st two cmdline options are -f input_file' then spin.cpp
will read the 'input_file' file, parse each line's options and
build a vector of vectors of cmdline options.
Then for each vector of cmdline options the worktyp tests are
run.
Also added a marker for the begin of a phase and end of phase.
Added the total performance and duration of the phase to the
marker.