Support drawing cpu_diagram for a phase
Currently the phase diagram can be drawn with the the values for
each metric.
Each metric on the diagram is computed from a bunch (29+) charts.
When the charts are all zoomed to a new interval (such as a phase)
then the metrics are recomputed but the values on the cpu_diagram
aren't updated until you hover over the diagram.
WHen you hover over the diagram the entire diagram is redrawn.
So I needed to have the diagram get updated whenever the all the
charts are zoomed to an interval.
I added code to handle the phase selected on the oppat cmdline.
For this to work:
1) all the charts have to be drawn (with no interval selected),
2) wait till the charts are drawn,
3) set the zoom interval to the selected phase,
   each chart looks for a change in the zoom interval and zooms
   itself
4) wait til the zoom is done (at which time the metrics will be
   recomputed)
5) redraw the cpu_diagram (and the table) with the updated metrics