Add macro define event feature and limiting events
Currently every event in the cpu_busy prf_obj gets a line in the
cpu_busy chart (above the cpu_occupancy bar).
If there are lots of events (lots of multiplexing and/or long
running jobs) then this can overwhelm the browser with 100s of MBs
of data.
Added a field 'dont_show_events_on_cpu_busy_if_samples_exceed' to
charts.json. If the number of samples is > this field then the
per-sample lines won't be drawn on the cpu_busy chart. The
cpu-occupancy lines (based on sched_switch or CSWITCH events) will
still be drawn. The flamegraphs come from this chart's event data
so the cpu-clock flamegraph won't be drawn (but the sched_switch
flamegraph will still be drawn).

Another added feature is a 'macro' event define feature in
charts.json.
Currently if you have an event+chart which is very similar to
another event+chart you have to duplicate the entry.
For instance there are 16 charts generated for the charts based on
event uops_dispatched_port.port_[0-7]. One set of 8 charts shows
the cycles/uop for each port and another set of 8 charts shows the
uops/sec for each port. The only difference is the port number.
This feature lets you define:
1) an event json object with a string (like %PORTNUM%)
2) the string to look for (like %PORTNUM%)
3) an array of strings to replace the 'look for' string.
   For instance ['port_0', 'port_1', 'port_2', etc]
A new event+chart object will be added for each new object.
This way I only have to have 1 base event+chart definition
instead of 8 almost duplicate definitions.