Handle new PORT strings and fix tot_line max value
When producing the haswell cpu_diagram table some of the chart
names are hard-coded in the js. I changed the name of some charts
and so I have to update the js.
Also fix an issue where the max value for the tot_lines wasn't
calculated properly (so that the top of some lines was cutoff).