Add more text fields to output and uops/nsec events
Currently the cpu_diagram has the generated event values with the
red/yellow/green light. But the user won't get a desc of the
value unless they hover of the value.
To fix this add a short desc to each value printed.
There are 2 ways to get a value:
1) the 'pfx' field will add the pfx string and then print the
values on the same line.
2) the 'tfld' object lets you add text at a specific x,y. This is
useful for text fields on different lines than the values.
Now there is a text field for each value.