Add display phase on hover/click and other changes
If phase data is passed to main.js then display the phase when
a user hover overs a point in the chart or clicks on a point in
the chart.
Also fix an issue where, if a chart is not displayed and a 'redraw
all' is requested, then the 'done with the redraw loop' never
finishes. See ck_if_need_to_set_zoom_redrawn_cntr().
Also add GUOPS_port_*_CHART charts to haswell table.