Add support for receiving long client data
Previously data from the client to the server was pretty small.
SO all the data would fit in one websocket frame.
Now I want to get back the image png for the cpu_diagram. The
image data is about 250 KB and is bigger than a frame.
So now I need to keep track of whether a msg is the 1st part,
a continuation and/or the end-of-message.
I save incomplete messages in a msg_data std::string variable.
This won't work well if there is more than 1 client per server.
But it seems to work well enough now.