Add phase and 'save png' features
I would like to be able to generate a series of PNGs files for a
given phase or for every X seconds between 2 phases.
So I need to be able to specify a starting and ending phase (the
--phase0 and --phase1 options) and an interval (the --ph_step_int
option) and an image base name to which the png images will be
saved (using the --ph_image option).
There is lots of checking that the phase0 and phase1 args are
found in the phase data.
And there is code to save the image data from the client.