Add 'xform=select_vars' feature, gen PNGs, & more
I'd like to be able to generate a move of how the cpu_diagram
changes over time. In particular I'd like to show the geekbench
data.
This will require specifying a start & time and a time step.
The data on the cpu_diagram comes from other charts. The
other charts will have to be recalculated for each subinterval.
Once that is done, the cpu_diagram will be drawn, then sent to
the backend. then do the next interval.
I can identify the geekbench phase start (and end) but the phases
have an idle time (much larger than the busy time in some cases).
This would lead to lots of idle images sent to the backend.
So I also look at the %busy chart and, if the system is less than
50% busy (where 100% busy means that 1 cpu is busy) then I skip
writing out that image.
Added feature 'xform:select_vars' to enable generating tot_line
for each variable on a chart.
To speed up generating the image/png, Each chart is zoomed to
the phase0 to phase1 time and the number of divisions on the
tot_line is set to
  divisions = (phase1_end_time - phase0_beg_time) / subinterval_tm
The tot_line data is then an array so I don't have to recompute
the all the charts for the cpu_diagram data, just lookup the
array values.