Changes for creating cpu_diagram movie
Changed header font size 20 pt so it would be more readable in
playback. All the other data is pretty small font (12 pt).
Added displaying the phase for the begin time and end time.
TO make things less confusing, I added code to find the nearest
phase if the beg tm or end tm is not within a phase.
And I added a 0.001 sec fuzz factor for checking if the time is
'within' the phase.
Also, I added whether the nearest time is 'bef begin' or
'after end' of the nearest phase.
Also added code to display with WikiChip.org copyright.