Add routine to create xlsx from json_table.json data
Currently the data for each update to the cpu_diagram is sent back
to the host into a json_table.json file.
The new routine in xls.cpp uses the libxlswriter library to create
an excel compatible xlsx file from the json table.
This is to enable checking the data and comparing/charting the
data.