Add --by_phase option to iterate over all phases
Currently we can iterate over phases by using a
--ph_step_int time_interval. So you can specify
--phase0 phs_beg --phase1 phs_end --ph_step_int 0.1
and oppat will start with phs_beg and gen a PNG each 0.1 seconds
until you get to the end of phs_beg.
This is mostly for geekbench where although we know the start/end
of each phase, we don't know when the idle time starts after the
active phase. Using the small steps I can check for the %busy
and skip the idle part of each phase.
For the spin benchmark I know the active phase of each subtest.
--phase0 phs_beg --phase1 phs_end --by_phase
will gen a png for phs_beg, then for the next phase, and so on
and stopping after phs_end.