Handle doing geekbench by_phase
Previously the geekbench PNGs were generated by stepping by 0.1
seconds through the benchmark and taking a PNG. If the system
usage drops below 50% (half of 1 cpu) for 2 consecutive intervals
then the PNGs is skipped for the rest of the phase... assuming
that the geekbench subtest has entered the idle part between
subtests.
Now I compute the cpu_busy usage on the initial load of the data
and use that to compute the end of each subtest.
Then I draw each subphase from the begin of the phase to the new
'begin of idle' part of the phase.
So now I get 1 PNG per subtest which includes all the non-idle
values.