Add --img_wxh_pxls option to override PNG size
Currently the PNG size (derived from the cpu_diagram canvas)
depends on the width of the screen. If you have the debug console
open then the width can vary.
This wouldn't be a problem except that the gen_mp4.py script has
hard coded offsets in it to draw boxes around sections of the PNG.
If the PNG size changes then the boxes and text positions are
messed up.
This lets you override the png (canvas) size.