Add CPI chart to block diagram
Most of the stats are 'per cycle' for the block diagram.
But the A53 is usually 1.0 IPC (instruction/cycle) at best.
It can retire 2 inst/cycle but it can be less than 1 due to cache
misses etc.
The CPI (cycles/inst) chart is just the inverse of the IPC chart.
The CPI chart makes it easier (IMHO) to see how far we might be
from the ideal 1 or 2 inst/cycle.