Add bat file to make PNGs for 'per phase'
Here is an example file for how to create 1 PNG for each phase
in the file.
The bat file will create files named 'pat0*.png' and pat_base.png.
The pat_base.png is the block diagram without any stats.
The pat00000.png is the 1st phase, pat000*1.png is the next phase,
etc.