Add a .py script to generate a gif from the PNGs
Here is a sample .py script which reads the pat0*.png files and
generates a .gif file from them.
Unfortunately the GIFs get really big since they are basically a
concatenation of the underlying PNGs. And we can't pause the GIF
playback.
These 2 reasonas are why I switched to using an MP4 to display
the phases.