Display labels for pie chart slices
Currently you can see what each pie slice corresponds to by
hovering over it.
But you can't hover over the slices for the PNGs.
Added a 'display legend' feature.
This is helpful but it can be hard (for me anyway) to match up
the color 'block' in the legend to the slice.
So now I add a slice number to the legend and, if there is enough
room in the slice to add the slice number to the pie slice, then
put the slice number in the pie slice.
This lets the user ID the bigger slices more easily.
Currently there are 2 types of blk_diag pie graphs: stalls and
L2_data_access_typ.
The stalls pie already has everything expressed in terms of
percentages.
The L2_data_access_typ pie has to compute the percentage from
the component graphs as a pct of the L2_data_access bytes/cycle
graph. So the 'total' chart is one of the component charts but
we don't want to show the chart in the legend.
The 'bal_slice' variable is used to ID the slice to be skipped
(not shown in the legend).