Add follow_proc option and tot_line data
Add oppat cmdline '--follow_proc string' option and tot_line
code for reporting follow_proc cpu usage.
I wanted to be able to see how much of the cpu a benchmark used.
Say, for geekbench, how much of each cpu does geekbench use.
There are a lot of 'per core' charts and stats on the block
diagram but I don't show a chart indicating if geekbench is the
main (or only) thing running on the core. Or if we are in a single
threaded phase or multi-core phase.
So I needed to be able to pass a string to oppat to indicate which
process we wanted to see a '%cpu' chart for.
If you enter '--follow_proc geekbench' (and you have geekbench
benchmark data and the cpu_diagram) then you get a pie chart
showing the %idle, %geekbench and %other for each core.