Add bar chart for block diagram
IPC is a central 'cpu resource utilization' stat. currently the
block diagram just lists the IPC per core with a 'stop light'.
The pie charts are useful but sometimes a bar chart is needed.
In particular, for IPC (instructions/cycle) where we know the max
is 2 IPC (for ARM cortex a53) then it is useful to show the IPC
per core relative to the max possible value of 2.0.
This commit adds a simple bar chart, with 1 vertical bar per core
and the height of the bar is proportional to the IPC.
Currently only 1 value per bar is supported but the bar chart
routine could be changed to do more than 1 value per bar.
Then I could replace the pie charts with a 'smaller' picture.
There also isn't any support yet for spaces between the bars.