Add lua script to read iostat.txt file
I tried collecting block and file IO stats with perf and the
amount of data interfered with perf hardware counter multiplexing.
So I installed iostat for ARM64 linux instead.
The output is a text file.
The timestamp in the text file is 2019-04-30T00:31:23-0400 format.
LUA can parse the date/time stamp and convert it epoch seconds.
This has to be converted to clk_mono time (which is
clock_gettime(CLOCK_MONOTONIC, &tp)).
So we pass the clocks1.txt file to the lua file as well.
clocks1.txt has the timestamp for clk_mono and gettimeofday (which
is epocs seconds (with decimal)).
Now we can convert epoch seconds to clk_mono.
We have to parse the text file and generate 1 record per timestamp
per device per variable ('per variable' means we have to have
different recs for rdKB/s and wrKB/s).
nicstat.lua and vmstat.lua are slightly better implementations of
how to read the text file and generate oppat data.