Add fields for OS diagram
As part of a top-down view of the system performance, one needs
the OS (operating system) view and then the chip view with the
cpu_diagram.
This commit adds the fields for a simple OS block diagram.
The OS diagram is added to the same canvas as the chip diagram.
The canvas is split into 'sectors' where sector 0 is the OS
and sector 1 is the chip. The x/y values in the chip diagram
are relative to the start of the cpu sector (so they don't have
to be redone).
Just drawing the boxes is not hard.
The OS boxes can be specified as a percentage so that everything
adjusts if I need to increase the Y pixels size.
The harder part is drawing the text values in the boxes.
Again relative x/y values are used... relative to the last box
drawn in the list.
There has to be a linkage between the chart data and the OS box
text field in which the chart data will be drawn.
The linkage is based on 'box_nm'.