Add OS diagram support, chart_options, sinebow clr
Currently there is support for a chip_diagram but to get a more
full system view, one needs to see the OS view.
See http://www.brendangregg.com/linuxperf.html or
http://www.brendangregg.com/Perf/linux_perf_tools_full.png
Added support for drawing OS diagram and populating the diagram
with values. This is a work in progress.
I use the flame graphs for the Application box,
The perf syscall event for the system calls (and VFS stats,
socket/tcp/ip stats). The cpu_busy chart is used for the scheduler
view.
Code was added to save the flamegraph and cpu_busy chart images
and paste them into the OS view. Currently the charts in the OS
view are just displayed with no support for displaying chart
values if you hover over the chart (whereas you do get popup info
if you hover over the cpu_busy or flamegraph charts themselves).
There is a big challenge for how to display the syscall data and
currently I limit is to the top 5 or so areas in the underlying
chart.
The 'syscall outstanding' chart is dominated by blocking calls
such as pselect6 or nanosleep. so the is code to omit the top
'blocking' events and display the next most outstanding events.
This list of 'events to skip over' should probably be in the
'.flds' file instead of hard coded in main.js.

The syscall event has maybe 60+ areas. This exhausted my color
list (I have a 10 color, a 20 color and a 40 color list).
So I added a sinebow color generator/selector which generates
up to 256 unique colors. Of course the colors can be pretty close
to each other but what can I do.

Also added code the read and use the chart_options passed from
charts.json. Most of the options are for the tot_lines and needed
for the syscall charts.
Based on https://gist.github.com/rhwlo/cd1b00763af2a415323d

Added code to use CH_TYPE_LINE/STACKED/BLOCK string instead of
just using 'line', 'stacked', 'block'. This makes it easier to
find code for line charts or stacked charts or block charts.