Add lua script to read trace_cmd syscall events
Each ftrace syscall area has an sys_enter_* event and a
sys_exit_* event. (so for the 'read' area there is a
sys_enter_read at the start of the read and a sys_exit_read at the
end of the event).
There isn't any marker to indicate a 'for certain' match of
sys_enter and sys_exit but I match on process name and thread
number. This seems to mostly eliminate false matches.
The duration is computed from the sys_exit_* - sys_enter_* time
stamp.
There is an routine for the charts.json 'syscall_all' event
and another routine for the sys_e.*read.* and sys_e.*write.*
events since I need to pull the bytes read/written from the arg
list.