Change lua_derived_tc_prf() args
Currently if the same lua file/routine is called for different
derived event, there isn't any way to for the routine to tell that
it is doing derived event 1 or derived event 2.
With this commit, we pass the derived event number to the lua rtn
so that (if needed) state tables can be built per derived event.
The syscall events ran into this issue.
I wanted to be able to generate a derived event of
'syscalls outstanding' and another derived event for
'syscalls outstanding non-blocking' (which excluded syscalls like
nanosleep, futex, wait, pselect etc.).
Without this commit, both calls to the syscall_all() lua rtn
would 'see' the sys_enter_X event but the first call for the
sys_exit_X event would 'see' the open sys_enter_X event (and emit
the derived event (and remove the open sys_enter_X event).
The second call to syscall_rtn() for the same sys_exit_X event
wouldn't see the open sys_enter_X.
This commit creates 'per derived event' input/output info to be
passed to the lua rtn and passes the derived event number to the
lua_rtn (so that the lua rtn can create 'per derived event number'
tables if needed.