Add more syscall charts
Add SYSCALL_OUTSTANDING_NONBLOCK_CHART,
SYSCALL_SOCKETS_ACTIVE_CHART.
Also added SHOW_EVEN_IF_ALL_ZERO to VMSTAT_MEM_cHART chart.
The SHOW_EVEN_IF_ALL_ZERO option is needed because the default
behavior is to not draw lines where the value is zero everywhere
and not add the entry to the chart legend.
This is an okay behavior where you have lots of lines such as
per proc/pid/thread lines which exist but don't have any cpu time.
But for some charts, you want to see that the line is there and
that its value is zero.
The VMSTAT_MEM_cHART is an example of this. The 'virtual memory
used' line should (usually?) be zero but the user won't know that
this line is all present and zero everywhere on the chart unless
we pass the SHOW_EVEN_IF_ALL_ZERO option for the chart.
This is a charts.json option: event_array->event->charts->options