Resend cpu_diagram .flds file every 'refresh'
Currently, if I make a change to a .flds file (such as
web\arm_cortex_a53_block_diagram.flds) and I want the changes to
show up in the browser, I have to quit opppat and restart it.
This requires re-reading all the data files, creating all the
chart data and sending the data to main.js.
The .flds file isn't processed (other than to read the file into
a string and send the string to main.js) so it is simple to
re-read the file when the browswer sends a 'Ready' websocket msg.
So I do that: reread the .flds file and send it every 'Ready' msg.
The 'Ready' msg gets sent every time you load 'localhost:8081' to
load the oppat page or hit refresh for an already loaded oppat
page.