Add evts_to_exclude support
For derived events and with regex event selection, one can select
hundreds of events for a chart. The syscalls trace event has maybe
500+ events.
The event to exclude feature allows one to select a range of
events (like with 'syscalls:sys_e.*') and then exclude some of the
selected events (like 'nanosleep').