Improve displaying OS_view chart info
Currently charts in the OS_view don't display more than the name
of the chart. And it is hard to add another chart.
This commit adds support for displaying the chart line data when
you hover over a line of the chart (just like for the underlying
chart from which the OS_view image of the chart was created).
Also, if you click on the chart then you jump to the underlying
chart.
The method of referencing a chart (in the .flds file) to be added
to the OS_view was also generalized to make it easier to add a
chart image and get the hover info.
The chart image size is much reduced compared to the underlying
chart so I have to allow for an increased 'compare size' for the
hover.
I made the hover logic a little more general so it can be called
more easily.