Add debugging for cpu_diagram overwrite issue
When creating a standalone html the cpu_diagram can overwrite
other parts of the diagram.
For example, the 'per cpu' pie charts legend can overwrite the
next charts.
The scaling of the added charts and text is not right.
As a workaround, I can hit 'cntrl -' (in the browser window to
zoom out) and, at 80%, the layout is what I expect.