Make sure only 1 phase matches begin/end string
The user can enter a string to select the phase begin and end
with the --phase0 and --phase1 options.
Currently if the phase string matches more than than one phase
then every matching string gets marked as the 'begin with this
phase'.
Add code to quit looking for matches after the first match.
Also, for phases while duration less than or equal to 0.0, don't
pass the phase to main.js.