Wait for OS_view charts before creating PNGs
When generating the PNGs, you have to wait till the cpu_diagram
text and charts are filled in... so you have to wait for all the
underlying charts to be drawn.
Now we've added the OS_view to the cpu_diagram. For the OS_view,
we take images of the underlying charts and draw the images into
the OS_view chart. So we have to now wait for the underlying
charts to be drawn, then for the chart to be converted to an image
then for the image to be ready and drawn into the OS_view.
This code (mostly) handles the 'wait for the underlying images to
be ready and drawn into the OS_view'.