Change script for OS_view and ARM specific blocks
The OS_view stuff on the PNG requires shifting the the existing
block and text stuff down by 800 pixels.
Some of the text on the cpu_diagram was also moved around.
I don't really have much in the way of info on the OS_view.
Just 1 slide.
Added the 'how to go directly to a sub-test (phase)' slide showing
the T=x for each phase.