Add test to avoid false GB 'in multi core' phase
There is code in oppat.cpp to try and identify geekbench phases.
The geekbench code doesn't emit timing info for the phases.
It does print a 'start X sub-test' message which the run_scripts
capture to identify a 'start of subtest' timestamp.
But there isn't a 'start of multi-core' message.
I download and read the gb.html results file to see if the test
has a multi-core section. Then for each subtest I look if more
than 1 cpu is kept busy and if more than 1.x cpus are busy for
more than a few milliseconds than I say the multicore phase has
started.
This works pretty well but sometimes would say the multicore stuff
has started when there isn't a multicore phase.
Now I check to see if geekbench reported a multicore phase for the
subtest before I look for the multicore phase.