Add temperature to the ARM CPU_diagram
At first I couldn't find the temperature on raspberry PI. It isn't
in the ftrace events (on ubuntu mate 64bit).
But it is in the /sys files.
So now add the temperature  chart to
web/arm_cortex_a53_block_diagram.flds so it shows up in the arm
cpu_diagram, the table and the xls file.