Change ARM script to match script I used for video
Not sure exactly how the script got out of sync but I copy the
script used to the output results dir and I see there are diffs.
Add syscalls event to trace-cmd
Split the export && iostat cmd to 2 lines. This is needed since
the returned pid is that of the shell rather than the iostat.
Also decrease the frequency of the perf stat cmd.