Add debug info and set event area (if avail)
The 'lst_ft_fmt_idx < 0' error message needed more debugging info.
I added info to display the charts.json event name and field
which had the err.
This error indicates that there wasn't a match on the event name
in the event table.
The reason for the error (in this case) was that the perf file had
only 1 ftrace event (sched_switch) and so perf omits an event id.
Also changed the code which constructs an 'event name with area'
(like sched:sched_switch) to check for and extract the event
area (if present in the event name).