Add c++ syscall derived event handler
There can be tons of syscalls ftrace events in a 5+ minute
geekbench run. Currently the events are processed by a lua
tc_syscalls.lua script.
The processing of the syscalls events is pretty straight forward
so I added a syscall_der_evt() routine to replace the lua script.
The drawbacks of the commit are that I have hard-coded a 'if lua
file == tc_syscalls.lua' then it calls the new c++ rtn.
Also hard coded is the offset of the syscalls IO 'count' field.
I have code elsewhere to check the offsets but I don't call it
here.
There should be a way in charts.json to say 'use syscall_der_evt()
and not the lua file/rtn' but I don't have that in place.