Add 'evt_use' field to charts.json
I wanted a way to not use events.
For instance, on haswell with syscalls ftrace event running
geekbench one can get 100's of thousands of events. It would be
helpful to have an easy way of dropping the event to see if the
slowdown in the chart generation time is due to the massive number
of syscall events. Also the gui is not very responsive.
So I added a 'evt_use' field to charts.json:
event_array->event->evt_use.
If src\rd_json.cpp sees: "evt_use":"n" then the event won't be
used. The event still has to be read from the data files (and then
skipped over) so there is still a slowdown during the read.
But the event isn't saved, any chart dependent on the event
doesn't get generated since there isn't any event data for the
chart.