Allow for creating gb_phase.tsv after run
Currently the script can only run right after the job finishes.
If there is a problem with creating gb_phase.tsv then you have to
rerun the job.
This change allows you to read the spin.txt file (which has all
the gb output (with each line prefixed by a timestamp and area)
and create gb_phase.tsv after the run.
Sample cmd lines:
 to read the spin.txt file and print the gb_phase.txt (but don't
 update it) afterwards do:
   gawk -v rd_md="1" -v odir="../oppat_data/lnx/gb10" -f \
   scripts\gb_rd_output2.gawk ..\oppat_data\lnx\gb10\spin.txt
 to read the spin.txt file and create the gb_phase.txt afterwards do:
   gawk -v rd_md="2" -v odir="../oppat_data/lnx/gb10" -f \
   scripts\gb_rd_output2.gawk ..\oppat_data\lnx\gb10\spin.txt

This allows me to fix issues with the parsing without having to
rerun the job.