Pass rd_md=0 to gawk script
scripts\gb_rd_output.gawk was changed to to allow creating
gb_phase.tsv after the run is complete.
Add the '-v rd_md="0"' args to the script to indicate we want to
create the gb_phase.tsv during the run (that is, while
run_perf_gb_haswell.sh is running).