Allow for --cpu_diagram filename to be .flds file
I use filename completion to generate the filename to be passed
to the --cpu_diagram option. The web\haswell_block_diagram.flds
filename comes before web\haswell_block_diagram.svg and I have,
so many times, passed the .flds filename instead of the .svg
filename.
This commit lets you pass either name to --cpu_diagram.
The code then changes the suffix to .svg for the svg file and
suffix to .flds for the flds file.