filter out perf process from syscalls event
The perf process calls read and write a ton.
It would be useful to have the perf events in syscalls but it
creates too many events and overloads the browswer for long
running geekbench4.
This commit uses the trace-cmd -f filter feature to exclude
the perf process name from the syscalls events.