Add more phase handling for geekbench4
Add support for geekbench4 idle periods between sub-tests.
geekbench 4 puts idle time betweeen each sub-test. I'd like to be
able to zoom in to just the non-idle part of a sub-test,
especially when generating PNGs.
So we insert an 'idle: X' subtest for each X sub-test.
The duration is initially 0.0. The idle phase end-time is set to
the non-idle end-time. The non-idle end-time is decreased by
0.00001 seconds (so that we can sort by end-time and get things
in the right order).
We look through the total cpu utilization where the utilization
is computed over N intervals.
If the FIND_IDLE_PHASE option is in the file_list.json file then
we look for the utilization to be < 40.0 (where on a 4 cpu system
then a 400.0 utilization means all 4 cpus 100% busy).
If util < 40.0 (in ck_phase_single_multi() ) then this marks the
begin of the idle part of the sub-test.
The 'idle: X' phase for sub-test X is then computed based on the
begin of the start time. The end-time for the non-idle sub-test
X is set to the start of the idle time.