Allow for more delay on first 'zoom to phase'
First the charts are drawn for the entire interval.
This can take a while depending on the amount of data.
Then (for mk_pngs_for_phase.bat) we zoom to each phase (or
non-idle phase for geekbench 4). But the first zoom can take a
while too. So leave the longer MyDelay interval until after the
first zoom is done.