remove FLD_TYP_DROP_BY_ONE_AFTER. add 'use' event
The FLD_TYP_DROP_BY_ONE_AFTER chart.json fld option wasn't ever
used and now it is removed.
I wanted a way to easily 'not use' some of the files in the
file_list.json file. So I added a "use":"n" field for the list of
files. The default is "use":"y" If "use":"n" is seen then that
line of the file_list.json file is skipped.