Optimize fill_data_table, drop DROP_BY_ONE_AFTER
In fill_data_table(), some work is done before checking if the
event is the one we want. Put the checking at the top of the loop.
This speeds things up.
Remove the code for DROP_BY_ONE_AFTER.
Add code to skip a line in the file_list.json if use_line != y.
This is the file_list.json "use":"y" (the default) option.