optimize prf_parse_text() a little
The routine looks for the perf event name on the text line.
This can be costly if you have lots of events in the file.
In particular, for a file with event groups, it seems that all the
events are in the same position... 60+ events in maybe 10 event
groups.
So to optimize, I handle the case where the events are in the same
position... save off the position of the last 'find' for event
and look for that field in the same position next time.
This is much quicker for the lots of events in event groups case.

Also, the order in which the events appear within an event group
is always the same. So I have code to try looking for 'the next
event' in the event list first. If the current event isn't the
guessed 'next event' then loop over all the events.
If these optimizations don't work then we turn off the
optimization. That is, if we loop over all the events checking if
any events is in the expected position, and we don't find any
match, then we disable the optimization.